+++ src/core/CMakeLists.txt
@@ -458,9 +458,8 @@ SET(QGIS_CORE_SRCS
   fieldformatter/qgsvaluemapfieldformatter.cpp
   fieldformatter/qgsvaluerelationfieldformatter.cpp
 
-  geocms/qgsgeocmsconnection.cpp
-  geocms/qgsgeonodeconnection.cpp
-  geocms/qgsgeonoderequest.cpp
+  geonode/qgsgeonodeconnection.cpp
+  geonode/qgsgeonoderequest.cpp
 
   ${CMAKE_CURRENT_BINARY_DIR}/qgscontexthelp_texts.cpp
   ${CMAKE_CURRENT_BINARY_DIR}/qgsexpression_texts.cpp
@@ -707,9 +706,8 @@ SET(QGIS_CORE_MOC_HDRS
 
   qgsuserprofilemanager.h
 
-  geocms/qgsgeocmsconnection.h
-  geocms/qgsgeonodeconnection.h
-  geocms/qgsgeonoderequest.h
+  geonode/qgsgeonodeconnection.h
+  geonode/qgsgeonoderequest.h
 )
 
 IF (NOT WITH_QTWEBKIT)
@@ -1083,9 +1081,8 @@ SET(QGIS_CORE_HDRS
   fieldformatter/qgsvaluemapfieldformatter.h
   fieldformatter/qgsvaluerelationfieldformatter.h
 
-  geocms/qgsgeocmsconnection.h
-  geocms/qgsgeonodeconnection.h
-  geocms/qgsgeonoderequest.h
+  geonode/qgsgeonodeconnection.h
+  geonode/qgsgeonoderequest.h
 
   metadata/qgslayermetadata.h
   metadata/qgslayermetadatavalidator.h
@@ -1109,7 +1106,6 @@ INCLUDE_DIRECTORIES(
   ${CMAKE_SOURCE_DIR}/src/core/expression
   fieldformatter
   geometry
-  geocms
   geonode
   layertree
   layout
